server/requirements.txt
django-cors-headers==3.0.2
django-extensions==2.1.9
django==2.2.2
djangorestframework==3.9.4
psycopg2==2.8.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
